LoRa receive from transmitter and acknowledge.
This work is using the lora_sx1276 driver from: https://github.com/belyalov/stm32-hal-libraries/blob/master/doc/lora.md
RX is performed by state machine and interrupt.
TX is performed by DMA.

The system works perfectly on the smaller transmitter project which is not included here. That simple project also uses this library but is bare metal.

This project is FreeRTOS and as soon as a send is initiated, the interrupt collapses. I have documented the issue here: https://github.com/belyalov/stm32-hal-libraries/blob/master/doc/lora.md

Although an easy library to implement for basic use, it is ageing and there is little support. I am checking in this work and moving to a better supported driver.